Minor fixed in stats/core.py
