+++ sora/stats/__init__.py
@@ -1,2 +1,5 @@
 from .core import Parameters
 from .optimize import OptimizeResult, least_squares, differential_evolution
+from .models import ellipse, ellipseError
+
+__all__ = ['Stats']
